separate bootstrap api client to its own mod
