+++ nym-vpn-core/crates/nym-vpn-api-client/src/lib.rs
@@ -6,10 +6,12 @@ pub mod types;
 
 pub(crate) mod jwt;
 
+mod bootstrap;
 mod client;
 mod error;
 mod request;
 mod routes;
 
-pub use client::{BootstrapVpnApiClient, VpnApiClient};
+pub use bootstrap::BootstrapVpnApiClient;
+pub use client::VpnApiClient;
 pub use error::{HttpClientError, VpnApiClientError};
